requirements.txt
Django==2.2.2
django-filter==2.1.0
djangorestframework==3.9.4
Markdown==3.1.1
psycopg2==2.7.3.2
pytz==2019.1
sqlparse==0.3.0
django-cors-headers==2.4.0
django-environ==0.4.5
requests==2.22.0
jsonfield==2.0.2
django-background-tasks==1.2.0
django-postgres-extensions==0.9.3
Faker==2.0.1
